rename MOE.py to variance_measures.py. also set z score back to .9, does not effect cv calculation and no good reason to change
